NetworkAI/requirements.txt
pandas>=1.3.0
matplotlib>=3.4.0
python-dotenv>=0.19.0
plotly>=5.3.0
openai>=1.0.0
groq>=0.4.0
tk>=8.6.0  # Usually comes with Python, but listing for documentation
